requirements.txt
fastapi
uvicorn
confluent-kafka
protobuf
sqlalchemy
python-dotenv
pydantic
asyncio
starlette
google
